Fix merge conflict error
