--- src/components/FirebaseImage.tsx
@@ -1,17 +1,0 @@
-import { getDownloadURL, ref } from "firebase/storage";
-import { useMemo, useState } from "react";
-import { ImageProps } from "react-bootstrap";
-import { storage } from "../config/firebase";
-import Image from "react-bootstrap/Image";
-
-export const FirebaseImage = ({ ...props }: ImageProps) => {
-  const [imageSrc, setImageSrc] = useState("");
-
-  useMemo(() => {
-    getDownloadURL(ref(storage, props.src)).then((val) => {
-      setImageSrc(val);
-    });
-  }, [props.src]);
-
-  return <Image {...props} src={imageSrc} />;
-};
